+++ rosomaxa/src/hyper/static_selective.rs
@@ -8,10 +8,14 @@ use std::sync::Arc;
 pub type HeuristicProbability<C, O, S> = (Box<dyn Fn(&C, &S) -> bool + Send + Sync>, PhantomData<O>);
 
 /// A type which specifies a group of multiple heuristic strategies with their probability.
-pub type HeuristicGroup<C, O, S> = Vec<(
-    Arc<dyn HeuristicOperator<Context = C, Objective = O, Solution = S> + Send + Sync>,
+pub type HeuristicSearchGroup<C, O, S> = Vec<(
+    Arc<dyn HeuristicSearchOperator<Context = C, Objective = O, Solution = S> + Send + Sync>,
     HeuristicProbability<C, O, S>,
 )>;
+
+/// A collection of heuristic diversify operators.
+pub type HeuristicDiversifyGroup<C, O, S> =
+    Vec<Arc<dyn HeuristicDiversifyOperator<Context = C, Objective = O, Solution = S> + Send + Sync>>;
 
 /// A simple hyper-heuristic which selects metaheuristic from the list with fixed (static) probabilities.
 pub struct StaticSelective<C, O, S>
@@ -20,7 +24,8 @@ where
     O: HeuristicObjective<Solution = S>,
     S: HeuristicSolution,
 {
-    heuristic_group: HeuristicGroup<C, O, S>,
+    search_group: HeuristicSearchGroup<C, O, S>,
+    diversify_group: HeuristicDiversifyGroup<C, O, S>,
 }
 
 impl<C, O, S> HyperHeuristic for StaticSelective<C, O, S>
@@ -37,7 +42,7 @@ where
         parallel_into_collect(solutions.iter().enumerate().collect(), |(idx, solution)| {
             heuristic_ctx.environment().parallelism.thread_pool_execute(idx, || {
                 unwrap_from_result(
-                    self.heuristic_group
+                    self.search_group
                         .iter()
                         .filter(|(_, (probability, _))| probability(heuristic_ctx, solution))
                         // NOTE not more than two search runs in a row
@@ -57,6 +62,10 @@ where
             })
         })
     }
+
+    fn diversify(&self, heuristic_ctx: &Self::Context, solutions: Vec<&Self::Solution>) -> Vec<Self::Solution> {
+        diversify_solutions(heuristic_ctx, solutions, self.diversify_group.as_slice())
+    }
 }
 
 impl<C, O, S> StaticSelective<C, O, S>
@@ -66,8 +75,11 @@ where
     S: HeuristicSolution,
 {
     /// Creates a new instance of `StaticSelective` heuristic.
-    pub fn new(heuristic_group: HeuristicGroup<C, O, S>) -> Self {
-        Self { heuristic_group }
+    pub fn new(search_group: HeuristicSearchGroup<C, O, S>, diversify_group: HeuristicDiversifyGroup<C, O, S>) -> Self {
+        assert!(!search_group.is_empty());
+        assert!(!diversify_group.is_empty());
+
+        Self { search_group, diversify_group }
     }
 }