+++ vrp-cli/src/extensions/solve/config.rs
@@ -547,7 +547,7 @@ fn configure_from_hyper(
                         .iter()
                         .map(|operator| create_operator(problem.clone(), environment.clone(), operator))
                         .collect::<Result<Vec<_>, _>>()?;
-                    get_static_heuristic_from_heuristic_group(heuristic_group)
+                    get_static_heuristic_from_heuristic_group(environment.clone(), heuristic_group)
                 } else {
                     get_static_heuristic(problem.clone(), environment.clone())
                 };
@@ -609,7 +609,7 @@ fn create_operator(
     problem: Arc<Problem>,
     environment: Arc<Environment>,
     operator: &SearchOperatorType,
-) -> Result<(TargetHeuristicOperator, TargetHeuristicProbability), String> {
+) -> Result<(TargetSearchOperator, TargetHeuristicProbability), String> {
     Ok(match operator {
         SearchOperatorType::RuinRecreate { probability, ruins, recreates } => {
             let ruin = Arc::new(WeightedRuin::new(